refactor: separate loader component file et al.
